==> R/04_describe_data.qmd <==
---
title: "04_describe_data"
format:
  html:
    embed-resources: true
knitr:
  opts_chunk:
    dev: "ragg_png"
    fig.path: "../results/"
editor: visual
---

# Load libraries

```{r}
#| label: setup
#| message: false
library("tidyverse")
library("patchwork")
```

# Describe data

In this document, we will make some descriptive statistics on the cleaned (transposed) data set.

-   Each row contains one sample

-   Each gene is a column

## Load data

```{r}
#| label: Read data
data <- read.table("../data/02_transposed_data.csv", header = TRUE, sep = "", 
                   check.names = FALSE)
```

## Number of genes

```{r}
#| label: Count number of genes
data |>
  select(-disease.state) |>
  ncol()
```

## Samples

Counting the sample size for each group (diabetic and none-diabetic) based on the column "disease.state".

```{r}
#| label: Count sample size for each group
data |>
  group_by(disease.state) |>
  summarise(samples.count = n(), .groups = "drop")
```

There are in total 14481 genes in both data sets.

There are 9 samples in the diabetic data set (T2D).

There are 54 in the control data set (none-diabetic).

## Statistics

The basic statistics for each group is calculated separately and displayed in a tibble.

The diabetic data set (T2D):

```{r}
#| label: Statistics for T2D data
stats_T2D <- data |>
  filter(disease.state == "T2D") |>
  select(-disease.state) |>
  # Calculations
  summarise(across(everything(), 
           list(mean = ~ mean(.x, na.rm = TRUE),
                median = ~ median(.x, na.rm = TRUE),
                sd = ~ sd(.x, na.rm =TRUE)))) |>
  # Adjusting the tibble format of the results
  pivot_longer(cols = everything(),
               names_to = c("Genes", ".value"),
               names_pattern = "^(.*)_(mean|median|sd)$")

stats_T2D
```

The control data set (none-diabetic):

```{r}
#| label: Statistics for control data
stats_control <- data |>
  filter(disease.state == "non-diabetic") |>
  select(-disease.state) |>
  # Calculations
  summarise(across(everything(), 
           list(mean = ~ mean(.x, na.rm = TRUE),
                median = ~ median(.x, na.rm = TRUE),
                sd = ~ sd(.x, na.rm =TRUE)))) |>
  # Adjusting the tibble format of the results
  pivot_longer(cols = everything(),
               names_to = c("Genes", ".value"),
               names_pattern = "^(.*)_(mean|median|sd)$")

stats_control
```

## Visualization of mean gene expression

First, we combine the two statistic data sets (stats_T2D and stats_control) in order to plot together.

```{r}
#| label: Combined data
combined_stats <- bind_rows(stats_T2D |>
                              mutate(group = "T2D"),
                            stats_control |>
                              mutate(group = "None-diabetic"))
```

Density plot of the mean gene expression:

```{r}
#| label: Combined density plot
combined_density_plot <- combined_stats |>
  ggplot(aes(x = mean, fill = group)) +
  geom_density(alpha = 0.5) +
  theme_minimal() + 
  theme(legend.position = "bottom") +
  labs(title = "Density plot of mean gene expression",
       x = "Mean gene expression",
       y = "Density",
       fill = "Group")

combined_density_plot
```

Boxplot of the mean gene expression:

```{r}
#| label: Combined boxplot
combined_boxplot <- combined_stats |>
  ggplot(aes(x = group, y = mean, fill = group)) +
  geom_boxplot() +
  theme_minimal() +
  labs(title = "Boxplot of mean gene expression",
       x = "Group",
       y = "Mean gene expression")

combined_boxplot
```

Side-by-side histograms of the mean gene expression:

```{r}
#| label: Combined histogram
combined_histogram <- ggplot(combined_stats, aes(x = mean, fill = group)) +
  geom_histogram(bins = 60,
                 color = "black") +
  facet_wrap(~group) +
  theme_minimal() +
  theme(legend.position = "bottom") +
  labs(title = "Distribution of mean gene expression",
       x = "Mean gene expression",
       fill = "Group")

combined_histogram
```

## Conclusion

Both data sets contain 14481 different genes but differ in sample size as there are 9 T2D samples and 54 control samples.

Generally, the distribution of the mean and median gene expression is similar between both data sets with some genes with a higher expression causing a slightly right-skewed distribution.

Further statistical analyses will be performed combined with a gene selection to determine whether the genes differ systematically and significantly between the groups.
