==> R/05_alt_analysis.qmd <==
---
title: "05_alt_analysis"
format:
  html:
    embed-resources: true
knitr:
  opts_chunk:
    dev: "svg"
    fig.path: "../results/"
editor: visual
---

## Load libraries and data

```{r}
#| message: false
library("tidyverse")
library("matrixStats")
```

## Load data

```{r}
mixed_data <- read.csv("../data/02_clean_data.csv")

mixed_data_clean <- mixed_data |>
  select(IDENTIFIER, starts_with("GSM"))

mixed_data_transposed <- as.data.frame(t(mixed_data_clean))
```

```{r}
mixed_data_transposed <- mixed_data_transposed |>
  rename_with(~ as.character(slice(mixed_data_transposed, 1) |>
  unlist())) |>
  slice(-1)
```

# Calculating standard parameters

For this alternate analysis we wish to firstly calculate a set of standard parameters (mean, sd and finally p-value) as these will be the key for the visualisations done later in this section.

```{r}
p_values = numeric(nrow(working_data))

for (i in 1:nrow(working_data)) {
  # Perform a t-test on the current gene for control vs. treated
  # Assuming your groups are "control" and "treated"
  ttest <- t.test(working_data[i, disease.state == "T2D"], working_data[i, disease.state == "non-diabetic"])
  
  # Store the p-value
  p_values[i] <- ttest$p.value
}
```
