==> R/07_correlations.qmd <==
---
title: "07 Correlations"
format:
  html:
    embed-resources: true
knitr:
  opts_chunk:
    dev: "ragg_png"
    fig.path: "../results/"
editor: visual
---

## Load Libraries

```{r}
#| label: setup
#| message: false
library("tidyverse")
```

## Load Data

Selected genes from the method using median and selected genes from the method using linear regression and the p-values are downloaded as well as the T2D data set.

```{r}
#| label: read files
median_genes <- read.csv("../data/05_selected_genes_median.csv")
pval_genes <- read.csv("../data/06_selected_genes_pvalue.csv")
df_T2D <- read.csv("../data/03_T2D_data.csv")
```

## Join gene data sets

Using inner join two find the unique values within the two gene-data sets and save it as a vector. These genes are expected to be important for the analysis, since they are found by both gene-selecting methods.

```{r}
#| label: Join sets of genes
median_genes <- median_genes |> 
  select(-X)
pval_genes <- pval_genes |> 
  select(-X)
selected_genes <- median_genes |> 
  inner_join(pval_genes, 
             by = c("IDENTIFIER" = "gene"))

selected_genes_vec <- selected_genes |> 
  pull(IDENTIFIER)
```

## Heat map

Create a heatmap of selected genes by extracting all gene expression level from the T2D-data set, compute correlation of each gene, and reshape the correlation matrix into a pivot-longer structure suitable as input for heatmap.

```{r}
#| label: heatmap
# Pairwise Correlation
cor_data_top_genes <- df_T2D |>
  filter(IDENTIFIER %in% selected_genes_vec) |>
  select(IDENTIFIER, starts_with("GSM")) |>
  column_to_rownames("IDENTIFIER") |>
  t() |>
  cor(use = "pairwise.complete.obs")

# Reshapedata
reshape_cor_top_genes <- cor_data_top_genes |>
  as.data.frame() |>
  rownames_to_column("Var1") |>
  pivot_longer(cols = -Var1, 
               names_to = "Var2", 
               values_to = "value")

# Heat Map
ggplot(reshape_cor_top_genes, 
       aes(Var1, 
           Var2, 
           fill = value)) +
  geom_tile() +
  scale_fill_gradient2(low = "blue", 
                       high = "red", 
                       mid = "white", 
                       midpoint = 0, 
                       limits = c(-1, 1),
                       name = "Correlation") +
  geom_text(aes(label = round(value, 2)), 
            color = "black", size = 2) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, 
                                   hjust = 1)) +
  labs(title = "Top 15 Diabetes-Related Genes",
       x = "", y = "") +
  coord_fixed()
```
